requirements.txt
interactions.py==4.5.1
tinydb==4.8.0
python-dotenv==1.0.1
